requirements.txt
asgiref==3.4.1
asttokens==2.0.5
colorama==0.4.4
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
executing==0.8.0
icecream==2.1.1
psycopg2-binary==2.9.1
Pygments==2.10.0
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
